# swagger.yaml
openapi: "3.0.0"
info:
  title: Swagger flow by Vandit
  description: I've use Swagger API for the best Documentation.
  version: 1.1.0
  contact:
   email: vanditbera@gmail.com
   url: "https://vanditbera.in/"

# servers:
#   - url: "https://localhost:4000/api/v1"
#     description: for localhost - secure
#   - url: "http://localhost:4000/api/v1"
#     description: for localhost - regular

servers:
  - url: "{protocol}://localhost:4000/api/{version}"
    description: for localhost
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
      version:
        enum:
          - v1
          - v2
        default: v1

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    BearerAuth:
      type: http
      scheme: bearer

paths:
  /test:
    get:
      tags:
        - Tests
      summary: returns a message for test
      responses:
        200:
          description: All Good and Success
          content:
            application/json:
              schema:
                type: string
                example: "My string"
        400:
          description: Bad request
        500:
          description: Internal server error
  /testobject:
    get:
      tags:
        - Object
      summary: returns a test objects
      responses:
        200:
          description: All Good and Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: number
        400:
          description: Bad request
        500:
          description: Internal server error
  /courses:
    get:
      tags:
        - Array
      summary: returns all the courses in array
      responses:
        200:
          description: All Good and Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    price:
                      type: number
        400:
          description: Bad request
        500:
          description: Internal server error
  /mycourses/{courseId}:
    get:
      tags:
        - Array
      summary: returns course based on req ID
      parameters:
        - name: courseId
          in: path
          required: true
          default: 3
          schema:
            type: string
      responses:
        200:
          description: All Good and Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: number
        400:
          description: Bad request
        500:
          description: Internal server error
  /addcourse:
    post:
      tags:
        - Array
      summary: adds a new course to existing courses
      consumes:
        - application/json
      produces:
        - application/json
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                price:
                  type: number
      responses:
        200:
          description: All Good and Success
          content:
            application/json:
              schema:
                type: boolean
        400:
          description: Bad request
        500:
          description: Internal server error
  /coursequery:
    get:
      tags:
        - Array
      summary: trying to learn about query
      parameters:
        - name: location
          in: query
          required: true
          schema:
            type: string
            enum: [dhoraji, rajkot, ahmedabad]
        - name: device
          in: query
          required: true
          schema:
            type: string
            enum: [web, mobile]
      responses:
        200:
          description: All Good and Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  location:
                    type: string
                  device:
                    type: string
        400:
          description: Bad request
        500:
          description: Internal server error
  /courseupload:
    post:
      tags:
        - Array
      summary: uploading course image
      parameters:
        - in: header
          name: auth
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                filename:
                  type: string
                  format: binary
      responses:
        200:
          description: All Good and Success
        400:
          description: Bad request
        500:
          description: Internal server error